feat: rename component to LaceApp and swap into production
- Rename LaceAppWithDesignSystem to LaceApp for cleaner naming
- Update all test references to use new component name
- Replace app/page.tsx to use new DaisyUI design system layout
- User now sees beautiful themed interface instead of gray styling
- Project loading and selection now works in production
- Responsive mobile/desktop sidebars active

BREAKING CHANGE: Main app now uses design system layout
Sessions/agents/chat functionality still pending (Tasks 4-6)
